refactor(coordinator): extract focus capture + capsule window into coordinator/capsule_focus.rs — behavior unchanged, 446 tests green
Pure move of 20 free functions (incl. cfg-variant focus/frontmost-app capture,
capsule window show/hide/position, emit_capsule) + CapsuleLayoutState, from
coordinator.rs into a new sibling submodule. Made pub(super); reached from the parent
and sibling submodules via 'use capsule_focus::*;'. DeferredAsrBridge intentionally
left in the parent (shared ASR-bridge type, not focus/capsule). No logic/signature/
public-API changes; coordinator.rs 4934→4298 lines.

cargo check: 0 errors, 87 warnings (unchanged) · clippy: 0 errors
cargo test --lib: 446 passed · tsc / npm run build: green
